Copying isn't really a sensible thing to do with these objects
So copy constructor and assignment operator are deleted. It's also a
pain to decide how these operations should behave, ie. do we setup a
second watcher watching the same stuff or just two references to the
same watches. The first is less painful but also kind of wierd, ie. why
would we want to be told twice about events?